pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-environmental-intelligence"
version = "0.1.0"
description = "Agentic AI POC combining DEFRA / ALB public APIs (personal exploration)"
authors = [{ name = "Steve Dickinson" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "langgraph",
  "langchain",
  "langchain-openai",
  "mcp[fastmcp]",
  "pydantic>=2.0.0",
  "pydantic-settings",
  "requests",
  "httpx",
  "fastapi",
  "uvicorn[standard]",
  "pymongo",
  "psycopg2-binary",
  "pgvector",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pre-commit",
  "types-requests"
]

[project.scripts]
defra-agent-run = "defra_agent.agent.main:run_once"

[tool.hatch.build.targets.wheel]
packages = ["src/defra_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]
exclude = ["infra"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
mypy_path = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]